grid/backtester_and_optimizer/source/backtester_rsi/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-futures==1.0.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
yahooquery==2.3.1
